+++ CHANGELOG.md
@@ -6,6 +6,12 @@ This file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## UNRELEASED
+
+### Changed:
+
+  - Add `findLast` method to both iterable types       
+
 ## [2.1.1] - 2022-05-05
 
 ### Changed: